refactor(api-service): move timing-safe helpers to dedicated shared module fixes NV-8031
