+++ apps/api/src/app/integrations/usecases/azure-setup-oauth-callback/azure-setup-oauth-callback.usecase.ts
@@ -7,7 +7,7 @@ import {
   AzureSetupStateData,
   GenerateAzureSetupOauthUrl,
 } from '../generate-azure-setup-oauth-url/generate-azure-setup-oauth-url.usecase';
-import { areHexDigestsEqual } from '../../../shared/helpers/is-valid-hmac';
+import { areHexDigestsEqual } from '../../../shared/helpers/timing-safe-equal';
 import { splitOAuthState } from '../generate-chat-oath-url/chat-oauth-state.util';
 import { AzureSetupOauthCallbackCommand } from './azure-setup-oauth-callback.command';
