+++ apps/api/src/app/shared/helpers/timing-safe-equal.ts
@@ -1,0 +1,27 @@
+import { timingSafeEqual } from 'node:crypto';
+
+export function areStringsEqual(expected: string, provided: string): boolean {
+  const expectedBuffer = Buffer.from(expected);
+  const providedBuffer = Buffer.from(provided);
+
+  if (expectedBuffer.length !== providedBuffer.length) {
+    return false;
+  }
+
+  return timingSafeEqual(expectedBuffer, providedBuffer);
+}
+
+export function areHexDigestsEqual(expected: string, provided: string): boolean {
+  if (expected.length !== provided.length) {
+    return false;
+  }
+
+  const expectedBuffer = Buffer.from(expected, 'hex');
+  const providedBuffer = Buffer.from(provided, 'hex');
+
+  if (expectedBuffer.length !== providedBuffer.length) {
+    return false;
+  }
+
+  return timingSafeEqual(expectedBuffer, providedBuffer);
+}
